upscaler-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
basicsr==1.4.2
realesrgan==0.3.0
pika==1.3.2
redis==5.0.1
